backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ruff==0.1.14
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
# Database
motor==2.5.1
pymongo==3.12.0

qdrant-client==1.7.0
# AI / ML
sentence-transformers==3.0.0
httpx==0.26.0
groq==0.4.0
# Parsing
python-docx==1.1.0
python-docx==1.1.0
pypdf==3.17.4
